move transactor to modal
